requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
transformers==4.30.2
torch==2.6.0
python-docx==0.8.11
requests==2.26.0
--only-binary :all:
tokenizers==0.13.3
accelerate==0.20.3 